gui/requirements.txt
pika==1.2.0
dataclasses-json==0.5.4
numpy==1.19.2
tensorflow==2.4.2
Pillow==8.2.0
